requirements.txt
psycopg[binary]
python-dotenv
httpx